+++ test/e2e/main/v2/fixtures/web-tutorial-modified/mocks.js
@@ -1,6 +1,5 @@
 /*
-Copyright 2019 Javier Brea
-Copyright 2019 XbyOrange
+Copyright 2021 Javier Brea
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -9,28 +8,24 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-"use strict";
-
-const { Core } = require("@mocks-server/core");
-const InquirerCli = require("../../../index");
-
-const handleError = (error) => {
-  console.error(`Error: ${error.message}`);
-  process.exitCode = 1;
-};
-
-const start = () => {
-  try {
-    const mocksServer = new Core({
-      plugins: [InquirerCli],
-    });
-
-    return mocksServer.start().catch(handleError);
-  } catch (error) {
-    return handleError(error);
-  }
-};
-
-module.exports = {
-  start,
-};
+module.exports = [
+  {
+    id: "base",
+    routesVariants: ["get-users:success", "get-user:1", "get-users-new:success"],
+  },
+  {
+    id: "user-2",
+    from: "base",
+    routesVariants: ["get-user:2"],
+  },
+  {
+    id: "user-real",
+    from: "base",
+    routesVariants: ["get-user:real"],
+  },
+  {
+    id: "users-error",
+    from: "base",
+    routesVariants: ["get-users:error"],
+  },
+];