+++ CHANGELOG.md
@@ -9,6 +9,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 ### Fixed
 ### Removed
+### BREAKING CHANGE
+
+## [2.0.0-beta.2] - 2021-02-16
+
+### Changed
+- chore(deps): Update mocks-server/core dependency. Adapt tests.
+
+## [2.0.0-beta.1] - 2021-02-14
+
+### Added
+- feat: Display current mock.
+- feat: Display menus for changing current mock.
+- feat: Add menus for changing route variant and restore variants
+
+### Changed
+- feat: Display legacy options and menus only when core `pathLegacy` setting has value. Add toggle legacy watch menu.
+- refactor: Refresh inquirer main options every time main menu is displayed.
+
+### Fixed
+- fix: Resolve previous inquirers before displaying a new one
+- fix: Start promise was never resolved
+
+### BREAKING CHANGE
+- Changed format of `cli` option to boolean (now `--no-cli` has to be used to disable it)
 
 ## [1.4.1] - 2020-12-25